requirements.txt
# LangChain Core
langchain
langchain-core

#Groq
langchain-groq

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#Data loaders
langchain_community
pypdf
arxiv 
pymupdf
bs4
wikipedia

#text-spliters
langchain-text-splitters

# HuggingFace 
sentence_transformers
langchain_huggingface

# for progress bar
tqdm

# Vector DB
faiss-cpu

#frontend
streamlit

#backend
fastapi
uvicorn
langserve
sse_starlette
